refactor(tarko): consolidate all search logic into unified SearchService
Unify scattered search processing logic across the web UI:
- Replace SearchResultNormalizer with comprehensive SearchService
- Remove redundant search handling from contentStandardizer
